refactor: extract providers and navigation from App
Move IAP and unsaved-changes into providers/, pull root stack
types/theme/FCM/bootstrap into navigation/, and leave App.tsx
as provider composition plus RootNavigator.
