+++ constants/screens.ts
@@ -1,0 +1,14 @@
+export const SCREENS = {
+  INTRO: "INTRO",
+  PERIOD_OVERVIEW: "PeriodOverview",
+  DAY_OVERVIEW: "DayOverview",
+  REGISTER: "Register",
+  LOGIN: "Login",
+  LOADING: "Loading",
+  SUMMARY: "Summary",
+  HOME: "Home",
+  PLANS: "Plans",
+  ALBUM: "Album",
+  DASHBOARD: "Dashboard",
+  PAYWALL: "Paywall",
+} as const;
