+++ hooks/useIAP.ts
@@ -1,9 +1,1 @@
-import { IAPProvider, useIAP as useIAPFromProvider } from "./useIAPProvider";
-
-// Re-export the provider for App.tsx
-export { IAPProvider };
-
-// Consumer hook for screens/components
-export const useIAP = () => {
-  return useIAPFromProvider();
-};
+export { useIAP } from "../providers/IAPProvider";
