+++ navigation/theme.ts
@@ -1,0 +1,11 @@
+import { DefaultTheme } from "@react-navigation/native";
+
+export const navigationTheme = {
+  ...DefaultTheme,
+  colors: {
+    ...DefaultTheme.colors,
+    text: "#292524",
+    background: "#ffffff",
+    card: "#ffffff",
+  },
+};
