+++ navigation/types.ts
@@ -1,0 +1,8 @@
+export type RootStackParamList = {
+  INTRO: undefined;
+  Register: undefined;
+  Login: undefined;
+  Loading: undefined;
+  Home: { screen?: string } | undefined;
+  Paywall: undefined;
+};
